requirements.txt
numpy>=1.19.0
scipy>=1.7.0
matplotlib>=3.3.0
PyYAML>=5.4.0
seaborn>=0.11.0
tqdm>=4.60.0